+++ src/app/[username]/import/page.tsx
@@ -1,0 +1,10 @@
+import ImportTab from "@/components/Import/ImportTab";
+import { Box } from "@mui/material";
+
+export default function ImportPage() {
+  return (
+    <Box m={2}>
+      <ImportTab />
+    </Box>
+  );
+}